+++ R/WaterSpec35.R
@@ -14,4 +14,5 @@
 #'   liyunmei at njnu.edu.cn
 #' @references Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
 #'   an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.
+#' @family Datasets
 NULL