+++ R/apply_to_image.R
@@ -84,7 +84,7 @@
 #' @references 
 #' Bi S, Li Y, Xu J, et al. Optical classification of inland waters based on
 #'   an improved Fuzzy C-Means method[J]. Optics Express, 2019, 27(24): 34838-34856.
-#'  
+#' @family Fuzzy cluster functions
 
 apply_to_image <- function(input, res,
                            output_image=TRUE, output_resultpng=FALSE, output_imRrs.n=FALSE,