+++ R/dt_water.R
@@ -1,0 +1,12 @@
+#' @title Absorption and scattering coefficient of pure water
+#' @description Absorption and scattering coefficient of pure water with unit m^-1
+#' @docType data
+#' @keywords datasets
+#' @name dt_water
+#' @usage dt_water
+#' @format A dataframe with 561 rows by 4 cols
+#' @details The wavelength is from 340 nm to 900 nm.
+#' @note Will add later.
+#' @references Will add later.
+#' @family Datasets
+NULL