+++ R/globalsVariables.R
@@ -18,4 +18,7 @@ utils::globalVariables(c('.','band','value','name','wv','variable',
                          'kmpp', 'imembrand','x','y','r','g','b',
                          'res.im.cluster','Chla','Rrs_clusters.default',
                          'wavelength.default','nm',
-                         'SRF_LIST'))
+                         'SRF_LIST',
+                         'dt_water',
+                         "Rrs443","Rrs490","Rrs510","Rrs560","Rrs620","Rrs665",
+                         "Rrs681","Rrs709","Rrs754","Rrs779"))